requirements.txt
fastapi>=0.100.0
uvicorn[standard]
sqlalchemy
pydantic>=2.0,<3.0
pydantic-settings
python-dotenv
mysql-connector-python
alembic
httpx
pytest
email-validator
python-docx
pymupdf
python-multipart
nltk
openai
langchain
langchain-community
requests
biopython
sentence-transformers
faiss-cpu==1.11.0
numpy<2.0
